requirements.txt
python-nmap==0.7.1
colorama==0.4.6
scapy==2.6.1
flask==3.1.2